requirements.txt
appnope==0.1.0
asn1crypto==0.23.0
awsebcli==3.12.1
bcrypt==3.1.4
blessed==1.14.2
boto3==1.4.7
botocore==1.7.47
cement==2.8.2
cffi==1.11.2
colorama==0.3.7
confusable-homoglyphs==2.0.2
cryptography==2.1.3
decorator==4.1.2
defusedxml==0.5.0
Django==1.11.7
django-allauth==0.34.0
django-cors-headers==2.1.0
django-extensions==1.9.7
django-google-maps==0.7.0
django-registration==2.3
django-rest-auth==0.9.2
django-storages==1.6.5
djangorestframework==3.7.3
docker==2.6.1
docker-py==1.7.2
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
jmespath==0.9.3
oauthlib==2.0.6
olefile==0.44
paramiko==2.4.0
parso==0.1.0
pathspec==0.5.0
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
pyasn1==0.3.7
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.0
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
PyYAML==3.12
requests==2.9.1
requests-oauthlib==0.8.0
s3transfer==0.1.11
semantic-version==2.5.0
simplegeneric==0.8.1
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
traitlets==4.3.2
typing==3.6.2
wcwidth==0.1.7
websocket-client==0.44.0
